+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Bug fixes
 
-* NPE problem happened in SharedRealm.finalize() (#3730).
+* NPE problem in SharedRealm.finalize() (#3730).
 * `RealmList.contains()` and `RealmResults.contains()` now correctly uses custom `equals()` methods on Realm model classes.
 * Build error when the project is using Kotlin (#4087).
 * Bug causing classes to be replaced by classes already in Gradle's classpath (#3568).